# specs/openapi.yaml
openapi: 3.0.3
info:
  title: AgentWallet Protocol API
  description: |
    Infrastructure for autonomous AI agents that pay their own way.
    
    Agents don't ask permission. They execute. This API accepts crypto payments
    (SOL) for services like voicemail processing, failure analytics, and verification.
    
    ## Authentication
    
    No API keys. No signup. Agents identify via `agent_id` and pay per-use with SOL.
    
    ## Payment Flow
    
    1. Check current pricing: `GET /api/pricing`
    2. Send SOL to service wallet
    3. Include transaction signature in request
    4. Service verifies on-chain and processes
    
    ## Free Tier
    
    Each `agent_id` gets 1 free voicemail to test the service.
    After that, payment is required.
    
  version: 0.1.0
  contact:
    name: Inksky
    url: https://inksky.net
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: https://inksky.net
    description: Production server
  - url: http://localhost:3000
    description: Local development

tags:
  - name: Core
    description: Essential service endpoints
  - name: Voicemail
    description: Voicemail processing service
  - name: Analytics
    description: Usage and health metrics

paths:
  /api/health:
    get:
      summary: Health check
      description: Returns current system health, pricing, and status
      tags: [Core]
      responses:
        '200':
          description: System healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '500':
          description: System degraded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/pricing:
    get:
      summary: Get current pricing
      description: |
        Returns current SOL/USD rate and service prices.
        Prices update every 5 minutes via CoinGecko.
      tags: [Core]
      responses:
        '200':
          description: Current pricing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PricingResponse'

  /api/status:
    get:
      summary: Public status page
      description: Detailed system status for monitoring
      tags: [Analytics]
      responses:
        '200':
          description: System status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'

  /api/agent/balance:
    get:
      summary: Check wallet balance
      description: Check SOL balance for a given wallet address
      tags: [Core]
      parameters:
        - name: wallet
          in: query
          required: true
          schema:
            type: string
            example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
      responses:
        '200':
          description: Balance information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '400':
          description: Invalid wallet address
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/voicemail/process:
    post:
      summary: Process voicemail
      description: |
        Transcribe voicemail audio and extract structured intent.
        
        **Free tier:** 1 voicemail per agent_id
        **Price:** 0.001 SOL (~$0.20 USD) per voicemail after free tier
        **Limit:** Maximum 2 minutes audio
        
        ## Payment
        
        After free tier, include payment object with Solana transaction signature:
        ```json
        {
          "payment": {
            "signature": "5xKp9..."
          }
        }
        ```
        
        Transaction must:
        - Be confirmed on Solana mainnet
        - Be less than 5 minutes old
        - Send >= required amount to service wallet
        - Not have been used before (24h deduplication)
      tags: [Voicemail]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VoicemailProcessRequest'
      responses:
        '202':
          description: Voicemail queued for processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VoicemailProcessResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '402':
          description: Payment required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaymentRequiredResponse'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/voicemail/status:
    get:
      summary: Check job status
      description: |
        Get status of a voicemail processing job.
        Results stored for 24 hours (polling fallback if webhook fails).
      tags: [Voicemail]
      parameters:
        - name: job_id
          in: query
          required: true
          schema:
            type: string
            example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          description: Job not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: "healthy"
        service:
          type: string
          example: "AgentWallet Protocol"
        version:
          type: string
          example: "0.1.0"
        pricing:
          type: object
          properties:
            sol_usd_rate:
              type: number
              example: 200.50
            prices:
              type: object
        uptime:
          type: object
        free_tier:
          type: object

    PricingResponse:
      type: object
      properties:
        service_wallet:
          type: string
          example: "7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU"
        sol_usd_rate:
          type: number
          example: 200.50
        timestamp:
          type: string
          format: date-time
        prices:
          type: object
          properties:
            voicemail:
              type: object
              properties:
                sol:
                  type: number
                  example: 0.001
                usd:
                  type: number
                  example: 0.20
            voicemail_priority:
              type: object
            log_failure:
              type: object
            kyc_verify:
              type: object
        free_tier:
          type: object
          properties:
            enabled:
              type: boolean
            limit:
              type: integer
              example: 1

    StatusResponse:
      type: object
      properties:
        service:
          type: string
        status:
          type: string
          enum: [operational, degraded]
        pricing:
          type: object
        infrastructure:
          type: object
          properties:
            solana:
              type: object
            webhooks:
              type: object
        service:
          type: object

    BalanceResponse:
      type: object
      properties:
        wallet:
          type: string
        balance_sol:
          type: number
          example: 0.05
        balance_usd_approx:
          type: string
          example: "$10.00"
        can_afford:
          type: object
          properties:
            voicemail_process:
              type: integer

    VoicemailProcessRequest:
      type: object
      required:
        - audio_url
        - webhook_url
        - agent_id
      properties:
        audio_url:
          type: string
          format: uri
          example: "https://storage.example.com/voicemail-123.mp3"
          description: URL to audio file (mp3, wav, m4a, ogg, webm)
        webhook_url:
          type: string
          format: uri
          example: "https://my-agent.com/webhooks/voicemail"
          description: HTTPS URL to receive results
        agent_id:
          type: string
          minLength: 3
          maxLength: 32
          pattern: '^[a-zA-Z0-9_-]+$'
          example: "my_agent_001"
          description: Unique identifier for this agent
        priority:
          type: boolean
          default: false
          description: Skip queue for faster processing (2x price)
        payment:
          type: object
          description: Required after free tier exhausted
          properties:
            signature:
              type: string
              example: "5xKp9..."
              description: Solana transaction signature

    VoicemailProcessResponse:
      type: object
      properties:
        job_id:
          type: string
          example: "550e8400-e29b-41d4-a716-446655440000"
        status:
          type: string
          example: "queued"
        free_tier:
          type: boolean
        remaining_free:
          type: integer
          example: 0
        charged:
          type: object
          nullable: true
          properties:
            sol:
              type: number
            usd:
              type: number
        payment_verified:
          type: boolean
        eta:
          type: string
          example: "30s"

    JobStatusResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
          enum: [queued, processing, completed, failed, unknown]
        result:
          type: object
          nullable: true
          properties:
            transcription:
              type: string
            intent:
              type: string
              enum: [callback_request, appointment_scheduling, information_request, complaint, sales_inquiry, other, unknown]
            callback_number:
              type: string
              nullable: true
            urgency:
              type: string
              enum: [high, medium, low]
            summary:
              type: string
            entities:
              type: object
            confidence:
              type: number
        error:
          type: string
          nullable: true

    PaymentRequiredResponse:
      type: object
      properties:
        error:
          type: string
          example: "PAYMENT_REQUIRED"
        message:
          type: string
        pricing:
          type: object
          properties:
            sol:
              type: number
            usd:
              type: number
            rate:
              type: number
        service_wallet:
          type: string
        free_tier_used:
          type: integer

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: string
          nullable: true